Update GetParamByID with null support
