+++ storage/storage.go
@@ -9,4 +9,5 @@ import (
 type Storage interface {
 	db.Querier
 	ListSummary(ctx context.Context, userID int64) ([]ListSummaryRow, error)
+	GetParamByID(ctx context.Context, userID int64, paramID int64) (GetParamByIDRow, error)
 }
